Avoid .yml  plugins file
When using the  file for loading all the available plugins, Electron seems can't find the relevant file due to path issue. So, to avoid the issue, it is best for now to load all the plugins directly into the configuration Command line interface to a user's defaults object. This is definitely a temporary solution for an ongoing work as there's must be a better alternative way to do this. Check the https://github.com/nklayman/vue-cli-plugin-electron-builder/issues/640 for further information